Revert (#11)
There was a bug with this PR where the Apache container was actually based on the Cypress image

This PR reverts that change, as well as the "corporate proxy" functionality, which seems out of place and beyond the scope of this sample project.

I have a feeling CI tests were passing erroneously for this project as they were timing out instead of being a correct exit 0.